+++ src/Symfony/Component/TypeInfo/Type/UnionType.php
@@ -57,4 +57,32 @@ final class UnionType extends Type
 
         return $string;
     }
+
+    /**
+     * BC Layer for Symfony\Component\PropertyInfo\Type.
+     *
+     * @internal
+     */
+    public function setCollection(bool $collection): void
+    {
+        parent::setCollection($collection);
+
+        foreach ($this->types as $t) {
+            $t->setCollection($collection);
+        }
+    }
+
+    /**
+     * BC Layer for Symfony\Component\PropertyInfo\Type.
+     *
+     * @internal
+     */
+    public function setNullable(bool $nullable): void
+    {
+        parent::setNullable($nullable);
+
+        foreach ($this->types as $t) {
+            $t->setNullable($nullable);
+        }
+    }
 }