+++ src/Symfony/Component/PropertyInfo/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Introduce `PropertyDocBlockExtractorInterface` to extract a property's doc block
  * Restrict access to `PhpStanExtractor` based on visibility
+ * Deprecate the `Type` class, use `Symfony\Component\TypeInfo\Type` class of `symfony/type-info` component instead
 
 6.4
 ---